Failures to find buildkit-qemu-$arch are properly reported
When the buildkit binaries are not available on the host, the docker
build log only contains a cryptic `0.252 exec /bin/sh: exec format
error` or similar.  The only way to learn about the problem is by
inspecting the journalctl logs and seeing `msg="exec: \"buildkit-qemu-arm\":
executable file not found in $PATH"`.  This change should more directly
percolate the error up the stack.

Signed-off-by: Christian Mesh <christianmesh1@gmail.com>