tests: enable integration test run on windows
The plan is to enable to skip all the failing tests on
windows so as to have the CI running on WS2022; and
then we will go fixing each tests and enabling them in
separate PRs. This is the baseline.

This will be a helpful approach as we bring parity
between windows and unix.

Signed-off-by: Anthony Nandaa <profnandaa@gmail.com>