Enable users to specify the User-Agent header
To allow specifying a custom user agent, the llb.HTTP now
has a new HTTPOption named UserAgent allowing the user to
specify a custom User-Agent header for the http request.

This is also exposed in the Dockerfile frontend with an
additional argument --user-agent=<user-agent> in the ADD
command.

In case no User-Agent header is set, the default BuildKit
User-Agent is now being used instead of the default Golang
one to be more consistent with other requests buildkit is
making (i.e. requesting images).

Signed-off-by: Matthias Gliwka <matthias@gliwka.eu>