hack: add docker-compose file for development
This adds a simple docker compose file for development. The development
version of buildkit will build with the dockerfile, set debug to true,
use a tcp address to be compatible with the remote driver, and enable
the debug image through delve to allow for remote debugging.

It also configures additional services for help with development. At
the moment, this only includes jaeger for collecting traces but can also
include additional services in the future for things like metrics
collection.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>