hack: add otel collector, prometheus, and grafana to compose file
This includes the otel collector and uses the otel collector to direct
traces and metrics to the correct services. This is intended to act as
more of what a pipeline might look like in a production environment for
the traces and metrics.

Traces are forwarded from the otel collector directly to jaeger similar
to how that was configured before.

Metrics are forwarded to the prometheus exporter which prometheus will
use to scrape the metrics. A grafana instance has been included to use
with the prometheus instance to visualize the data. The default username
and password is `moby/moby` and the URL for the prometheus instance is
`http://prometheus:9090`. This state along with any dashboards will be
retained in your local development environment through a volume.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>