cniprovider: support `rootlesskit --detach-netns` (RootlessKit v2)
RootlessKit v2 added the support for `--detach-netns` mode.

This leaves containerd and buildkitd in the host netns so as to:
- Eliminate the slirp overhead for pull/push
- Allow accessing the real localhost registry.

Since rootless users has no `CAP_NET_ADMIN` in the host netns,
CNI plugins have to be executed in the "detached" netns (`$ROOTLESSKIT_STATE_DIR/netns`)
that is associated with slirp.

ref:
- rootless-containers/rootlesskit PR 379
- containerd/nerdctl PR 2723

Signed-off-by: Akihiro Suda <akihiro.suda.cz@hco.ntt.co.jp>