gateway: ensure that all process io is correctly closed
When closing pipes, we only need to close one side (which was previously
correct), but we do need to ensure that we correctly discard both sides
from our internal maps (which wasn't being done).

This could cause ExecProcess to potentially hang indefinitely. This was
originally detected because this hang meant that shutting down the GRPC
server with a graceful shutdown would wait for all ongoing RPCs to stop
- which would then never happen, and cause a deadlock.

Signed-off-by: Justin Chadwell <me@jedevc.com>