fix(docs): add the containers feature enabling step
By default, a fresh Windows installation does not come
with `containers` feature enabled since not all users
usually need the feature. It is therefore provided as
a _feature on demand (FoD)_.

This commit introduces the one-step powershell
command that enables this feature. This is necessary
for anything to run successfully.

Signed-off-by: Anthony Nandaa <profnandaa@gmail.com>