client: expose a method to create a trace client to replace the delegated tracer
This change removes the delegated tracer and replaces it with exposing
a method to create the trace client.

The primary benefit is it separates a client functionality from server
functionality. The delegation was only used by clients and it would
often happen automatically. If you initialized a client, it would
potentially report the same traces multiple times.

For the detect functionality, it was possible to turn off delegation on
the clients by enabling tracing or metrics from there. This prevents the
detect functionality from conflicting with the client delegation so it
always works properly now.

It also refactors the delegated span exporter a bit so its usage is not
tied to the `detect.Exporter()` method. It also only delegates to a
single span exporter which should help prevent accidentally sending
traces multiple times to the same client.

Clients are expected to use the new method and register a separate span
processor rather than use the tracer delegate.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>