tracing: refactor the trace recorder
The trace recorder is now a separate entity and does not wrap another
span exporter. It is added as another span processor to the underlying
tracer provider so the two can be disconnected from each other and this
removes the need to use `detect.Exporter()` to find the TraceRecorder
along with the need to invoke flush manually.

The trace recorder is wrapped with a simple span processor instead of
the batch one. That makes the flushing irrelevant for this purpose.

The trace recorder itself is also modified to avoid leaking goroutines
and respecting context cancellations.

This is part of a general refactor of the detect package to separate the
buildkit-specific functionality from the external exporter detection.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>