gha: add support for batch checking existing keys
In order to avoid getting rate limited when checking
if previous cache already exists use the Github REST API
that supports listing active keys.

This means additional GITHUB_TOKEN access is also
needed in addition to the RUNTIME_TOKEN needed for the
cache API. The implementation is done with a fallback -
if GITHUB_TOKEN is passed then it is used, if not then
the old method checking individual cache records is still used.

Signed-off-by: Tonis Tiigi <tonistiigi@gmail.com>