inline: fix uncompressed digest importing from multiple providers
When Moby adapters are used without containerd backend, the
descriptors are lacking the original compressed layer digest and
there is a workaround to link in the uncompressed digest. If
a build exporting the cache also imported it from the registry
then the original imported records with compressed digest needs
to be exported as well. If there were another exported records
that referred to the same layers via the uncompressed digest
through parent layers then such records could be overwritten
and the original provider from cache import could get lost.

Signed-off-by: Tonis Tiigi <tonistiigi@gmail.com>