remotecache: allow exporting root cache records
Allow exporting layers of root sources, like base images to
remote cache. Normally, this is not done as it is considered
wasteful as pulling cache vs. pulling source image again
does not have benefits.

Skipping base images in cache has other side effects though,
for example provenance of a cached build that is on top of
previous base image does not list the layers for the
LLB image step because that step did not run (because its
children were fully cached) and the image step itself
does not have a layer chain associated with it in the cache.

Signed-off-by: Tonis Tiigi <tonistiigi@gmail.com>