docs: dockerfile reference title in front matter
Use front matter to specify the doc title for the Dockerfile reference
in order to make the title programmatically accessible with Hugo, making
it easier to properly render the title in templates.

Specifying the title in the document directly means we have no control
over the rendering of the top-level heading.

Signed-off-by: David Karlsson <35727626+dvdksn@users.noreply.github.com>