linter: add lint rule for required json arguments
This lint requires json arguments to be used when the `SHELL` command
isn't used previously for `ENTRYPOINT` and `CMD`. This is because using
non-json arguments can block signals from properly being handled when
used in shell mode.

This uses the `Shell` attribute on the image configuration and therefore
requires the stage to be included in the build for it to properly run.
This is done because we don't want to force image config resolution on
every stage even if the stage isn't part of the build, but the only way
to definitively know if a custom shell is set is to look at the image
configuration.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>