buildkitd: configure telemetry providers with the global handlers
The global handlers are used as defaults for a bunch of OTEL functions
and we'll likely use it that way in the future too. While it's always
helpful to have a way to manually set the providers (and not rely on the
global providers), setting the global providers from the main function
is generally the correct thing to do.

This also removes the need for us to use the `WithTracerProvider` and
`WithMeterProvider` option to the grpc stats handler as it will use the
global configuration if this option isn't set. Since this code is in the
main function too, it's safe to rely on that behavior.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>