dockerfile/linter: check for nil linter in linter functions
During parsing, the linter can be nil and linter rules won't succeed at
running due to a nil dereference. Instead of checking for nil
everywhere, this modifies the linter to automatically disable itself if
the linter is nil.

Fixes a nil dereference panic happening when parsing `ENV` and `LABEL`
commands without a linter introduced by
6cfa4599029db7f2e6e83feaaa33984785ddd147.

Co-authored-by: Paweł Gronowski <pawel.gronowski@docker.com>
Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>