dockerfile: only report legacy key/value format when stage is reachable
This modifies the linter to only report the legacy key/value format if
the stage is reachable to be more compatible with the other lints that
operate this way.

Instead of reporting the warning in the parser, it records whether a
delimiter is present in the KeyValuePair and this gets checked when the
dispatch happens.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>