llb: convert envlist from slice to linked list
The semantics of envlist is that it gets many adds, and reads
but reads often only appear on a specific state of the list, usually
when all insertions have been completed.

Previous implementation did a full copy on each insert. For lookups
the options were to do a O(N) lookup per key or copy to a string
slice and then externally try to convert it to map with another copy.
In new version, the insertions are cheap and on first access optimized
map table is created that is then reused for next lookups.

Signed-off-by: Tonis Tiigi <tonistiigi@gmail.com>