fix: dot path normalized correctly for COPY
In my previous fix on #4825, I had removed this line
knowing that all that had been addressed in `pahtRelativeToWorkingDir`:

```go
	if cfg.params.DestPath == "." // <-- this one
		|| cfg.params.DestPath == "" ||
		cfg.params.DestPath[len(cfg.params.DestPath)-1] == filepath.Separator {
		dest += string(filepath.Separator)
	}
```

However, I had overlooked the `"."` and `""` scenarios. `""`.
The  `"/"`case is handled correctly in `system.NormalizePath()`.

This change therefore undoes this, to make sure "." is transformed
correctly to "./" during COPY operation, same to "" -> "./". This
is important especially for WORKDIR that are not `/`, so that
`COPY --link` operations are handled properly.

fixes #5070

Signed-off-by: Anthony Nandaa <profnandaa@gmail.com>