golangci: forbid uses of platforms.DefaultString()
The `platforms.DefaultString()` utility returns a platform-string for the
local platform. Starting with [containerd/platforms@38a74d2], a new format
is introduced to allow including OS-version, which is relevant for Windows.

Versions of `platforms.DefaultString()` after [containerd/platforms@38a74d2]
include the OS-version by default, which can be problematic if that string
is either parsed by code not yet updated to support the new format, or
(in general) when sending the platform to a remote / non-matching host
(e.g. a Windows client sending to a Linux machine).

Disallow using the `platforms.DefaultString()`, and recommend using
`platforms.Format(platforms.DefaultSpec())` instead (which produces
the old format without OS-version). Also add a warning about the
cross-platform risks of using this function.

[containerd/platforms@38a74d2]: https://github.com/containerd/platforms/commit/38a74d209d3bd4091fa83db35061ce32da31b5c3

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>