Restore env replacement process of ARG's variable names
Up to 1.7.1, variable names in ARG command could also use environment
replacement, but in commit 47a52a17b remove the replacement process, so
the replacement for variable names is no longer available in 1.8.0.
This PR restore this.
In addition, remove the Expand method of ArgCommand because it was no
longer used.

Signed-off-by: Mitsuru Kariya <mitsuru.kariya@nttdata.com>