Revisit secret mounting as environment variable:   * mount the secret as a file (existing behavior)   * additionally, mount the secret as an environment variable
'env' flag is used both as a boolean toggle (without value) to inject
the secret as an environment variable using the value of 'id' as a
default.
If provided a value ('env=MY_SECRET'), the secret will be injected into
the environment variable with the specified name.

In either case, the secret is always mounted as a file using prior
behavior.

Signed-off-by: a-palchikov <deemok@gmail.com>