remotecache: handle not implemented error for Info()
Moby graphdriver backend does not implement Info()
and such case should not be used as a signal for
blob needed to be skipped as it is unavailable.

This requires a change in Moby side as well to return
a typed error instead of string.

Signed-off-by: Tonis Tiigi <tonistiigi@gmail.com>
(cherry picked from commit 8fdf84f82181c897e10fd4c83b10adae6781c81f)